+++ source/Account.h
@@ -54,7 +54,7 @@ public:
 	const std::vector<Mortgage> &Mortgages() const;
 	void AddMortgage(int64_t principal);
 	void AddFine(int64_t amount);
-	void AddCrewLifeInsuranceCounter(int64_t deaths);
+	void AddCrewLifeInsuranceCounter(int deaths);
 	void DecayCrewLifeInsuranceCounter();
 	double CrewLifeInsuranceSalariesMultiplier() const;
 	int64_t CalculateCrewLifeInsuranceCost(int64_t salaries) const;
@@ -80,7 +80,7 @@ private:
 	int64_t maintenanceDue = 0;
 	// Your credit score determines the interest rate on your mortgages.
 	int creditScore = 400;
-	int64_t crewLifeInsuranceCounter = 0;
+	int crewLifeInsuranceCounter = 0;
 
 	std::vector<Mortgage> mortgages;
